commit based on changes made to get things compiling properly on Mac OS X most of the changes made were in regard to oversights in template composition also had to resolve quite a few cases where we included windows.h and other windows specific things opengl is also in a different location on macs, had to use the preprocessor definition __APPLE__ to properly include the gl libs

+++ branches/zapp/code/src/console/zonsole.cpp
@@ -89,7 +89,7 @@ bool Zonsole::handleKeyDown(const EventArgs &e)
 
 void Zonsole::chooseAuto(bool dir)
 {
-   size_t cnt = autoCompleteWnd->getItemCount();
+   int cnt = (int)autoCompleteWnd->getItemCount();
    int osel = getSelectedIndex(), nsel = -1;
 
    if (osel >= 0)
